+++ launch/robot_launch.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 # SPDX-FileCopyrightText: 1996-2023 Cyberbotics Ltd.
 # SPDX-FileCopyrightText: 2026 Dmytro Varich
 # SPDX-License-Identifier: Apache-2.0