+++ utils/check_map_values.py
@@ -1,5 +1,8 @@
 #!/usr/bin/env python3
 
+# SPDX-FileCopyrightText: 2026 Dmytro Varich
+# SPDX-License-Identifier: Apache-2.0
+
 import rclpy
 from rclpy.node import Node
 from nav_msgs.msg import OccupancyGrid